Added license
gdi developers, adding a license is a very important part of such a
large project, and we basically already have MIT as the standard for
HTwins STEM projects. Just add a freaking license, all it takes is 1
commit or even **PUSHING A SINGLE BUTTON** when creating the project in
order to have github ADD A LICENSE AUTOMATICALLY.

seriously, I don't know how you consistently fail at this, but add a
license. This software was meant for the community, not for us, so may
as well dedicate it as such with a share-alike.

Signed-off-by: ZweiSpeedruns <zcm727@gmail.com>